requirements.txt
pytest
django==5.0
djangorestframework
psycopg==3.2.6
django-extensions
pillow
django-cors-headers
requests
python-dotenv
django-eventstream[drf]==5.1.0
daphne==4.1.1
cryptography==3.4.6
channels==4.0.0